% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPalette.R
\name{getPalette}
\alias{getPalette}
\title{Get Palette from CSV Files in Directory}
\usage{
getPalette(
  aesthetic,
  colors,
  displayNames = FALSE,
  fileSuffix = "-colors",
  fileExtension = ".csv",
  directoryOfPalette = "extdata",
  thisPackageName = "styles.cmac"
)
}
\arguments{
\item{aesthetic}{The name of the aesthetic}

\item{colors}{A vector of color names}

\item{displayNames}{Logical, if TRUE display names of the colors}

\item{fileSuffix}{A suffix to be added to the file name (default: "-colors")}

\item{fileExtension}{The extension of the file (default: ".csv")}

\item{directoryOfPalette}{The directory of the palette data (default: "extdata")}

\item{thisPackageName}{The name of the package (default: "styles.cmac")}
}
\value{
The function returns a named list representing a color palette
}
\description{
This function reads palette data from CSV files located in a specified directory
}
