% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimesNewRomanFamily.R
\name{getTimesNewRomanFamily}
\alias{getTimesNewRomanFamily}
\title{Get Times New Roman Family}
\usage{
getTimesNewRomanFamily()
}
\value{
The function returns a font family as a string.
}
\description{
Gets the times new roman family equivielent across systems
}
\examples{
getTimesNewRomanFamily()
}
