% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cmac.R
\name{scale_cmac}
\alias{scale_cmac}
\title{Function that returns palettes or individual colors
This is done separately just to reduce noise in the file
Mainly, this just either displays the names of the colors or not when returing
It also allows you to return 1 or all of the colors}
\usage{
scale_cmac(aesthetic = NULL, ..., displayNames = FALSE)
}
\arguments{
\item{aesthetic}{Color aesthetic in 'fill', 'color', 'base', 'text','border','background', or 'grayedOut'}

\item{...}{single or vector of colors}

\item{displayNames}{TRUE if you want to see the names of each color. Default FALSE.}
}
\value{
hex codes of a subsetted CMAC palette
}
\description{
Function that returns palettes or individual colors
This is done separately just to reduce noise in the file
Mainly, this just either displays the names of the colors or not when returing
It also allows you to return 1 or all of the colors
}
\examples{
library(styles.cmac)

# Single valued colors: Basic elements
scale_cmac('text') # Text (a dark gray)

# Palettes: Used in scale fill or scale color below
scale_cmac('color', displayNames = TRUE) # All accent colors (line colors)
scale_cmac('fill',  displayNames = TRUE) # All main colors (fills)
scale_cmac('fill', 3)                    # 3 fills
scale_cmac('fill', 'red', displayNames = TRUE) # Just a red fill

}
\seealso{
\code{\link{theme_cmac}}

\code{\link{scale_color_cmac}}

\code{\link{scale_fill_cmac}}
}
