% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_cmac.R
\name{scale_color_cmac}
\alias{scale_color_cmac}
\title{CMAC ggplot Color Scale}
\usage{
scale_color_cmac(overrideWithFill = FALSE)
}
\arguments{
\item{overrideWithFill}{TRUE if you want to use the fill palette for the color mapping. (Default FALSE)}
}
\value{
ggplot color scale
}
\description{
CMAC ggplot Color Scale
}
\examples{
library(ggplot2)
library(styles.cmac)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=wt, color=as.factor(cyl))) +
    geom_point(size = 3, alpha = 0.5)

# Add the color
ggplotObject + scale_color_cmac()
}
