% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_cmac.R
\name{scale_fill_cmac}
\alias{scale_fill_cmac}
\title{CMAC ggplot Fill Scale}
\usage{
scale_fill_cmac(overrideWithAccent = FALSE)
}
\arguments{
\item{overrideWithAccent}{TRUE if you want to use the color palette for the fill mapping. (Default FALSE)}
}
\value{
ggplot fill scale
}
\description{
CMAC ggplot Fill Scale
}
\examples{
library(ggplot2)
library(styles.cmac)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=wt, fill=as.factor(cyl))) +
    geom_point(size = 3, alpha = 0.5)

# Add the fill color
ggplotObject + scale_fill_cmac()
}
