% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cmac.R
\name{theme_cmac}
\alias{theme_cmac}
\title{Create a CMAC Theme for ggplot}
\usage{
theme_cmac(inTimesNewRoman = FALSE)
}
\arguments{
\item{inTimesNewRoman}{TRUE if plot in Times New Roman Font (Default), FALSE uses ggplot main font. Only works on Windows.}
}
\value{
The function returns a ggplot theme
}
\description{
This function creates a custom theme for ggplot with several theme modifications
}
\examples{
library(ggplot2)
library(styles.cmac)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=wt, color=as.factor(cyl))) +
    geom_point(size = 3, alpha = 0.5) +
    scale_color_cmac()

# Add the theme
ggplotObject + theme_cmac()

}
\seealso{
\code{\link{scale_cmac}}

\code{\link{scale_color_cmac}}

\code{\link{scale_fill_cmac}}
}
