% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_cmac.R
\name{scale_color_cmac}
\alias{scale_color_cmac}
\title{CMAC ggplot Color Scale}
\usage{
scale_color_cmac(
  overrideWithFill = FALSE,
  colorOffset = 0,
  reverseOrder = FALSE,
  ...
)
}
\arguments{
\item{overrideWithFill}{TRUE if you want to use the fill palette for the color mapping. (Default FALSE)}

\item{colorOffset}{Offsets the ordering of the palette. E.g., colorOffset = 2 would move the first two colors to the end of the palette, starting at the 3rd color.}

\item{reverseOrder}{TRUE reverses the default order of the palette. Default FALSE. Applied after colorOffset.}

\item{...}{Additional parameters passed to either scale_color_manual or scale_fill_manual functions, such as guide = 'none'}
}
\value{
ggplot color scale
}
\description{
CMAC ggplot Color Scale
}
\examples{
library(ggplot2)
library(styles.cmac)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=wt, color=as.factor(cyl))) +
    geom_point(size = 3, alpha = 0.5)

# Add the color
ggplotObject + scale_color_cmac()

# Add the fill
ggplotObject + scale_color_cmac(overrideWithFill = TRUE)

# Add the color with no legend using an optional parameter
ggplotObject + scale_color_cmac(guide = 'none')

# Offset the colors by 1
ggplotObject + scale_color_cmac(colorOffset = 1)

# reverse the order of the palette
ggplotObject + scale_color_cmac(reverseOrder = TRUE)

}
\seealso{
\code{\link{theme_cmac}}

\code{\link{scale_cmac}}

\code{\link{scale_fill_cmac}}
}
