% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_cmac.R
\name{scale_fill_cmac}
\alias{scale_fill_cmac}
\title{CMAC ggplot Fill Scale}
\usage{
scale_fill_cmac(overrideWithAccent = FALSE, ...)
}
\arguments{
\item{overrideWithAccent}{TRUE if you want to use the color palette for the fill mapping. (Default FALSE)}

\item{...}{Additional parameters passed to either scale_color_manual or scale_fill_manual functions, such as guide = 'none'}
}
\value{
ggplot fill scale
}
\description{
CMAC ggplot Fill Scale
}
\examples{
library(ggplot2)
library(styles.cmac)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=as.factor(cyl), fill=as.factor(cyl))) +

    # Note that gray border used for demonstration. Please use scale_color_cmac()
    # as much as possible
    geom_boxplot(size = 0.8, color = scale_cmac('gray', 'gray3'))

# Add the fill color
ggplotObject + scale_fill_cmac()

# Add the accent color
ggplotObject + scale_fill_cmac(overrideWithAccent = TRUE)

# Add the fill color and turn off legend with optional parameter
ggplotObject + scale_fill_cmac(guide = 'none')

}
\seealso{
\code{\link{theme_cmac}}

\code{\link{scale_cmac}}

\code{\link{scale_color_cmac}}
}
