% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cmac.R
\name{theme_cmac}
\alias{theme_cmac}
\title{Create a CMAC Theme for ggplot}
\usage{
theme_cmac(inTimesNewRoman = FALSE, removeTopAndRightBorder = TRUE, ...)
}
\arguments{
\item{inTimesNewRoman}{TRUE if plot in Times New Roman Font (Default), FALSE uses ggplot main font. Only works on Windows.}

\item{removeTopAndRightBorder}{TRUE if remove right and top parts of border. Recomended FALSE if using small multiples.}

\item{...}{Additional parameters that can be passed into the theme() function. Added last to theme.}
}
\value{
The function returns a ggplot theme
}
\description{
This function creates a custom theme for ggplot with several theme modifications
}
\examples{
library(ggplot2)
library(styles.cmac)

# Create basic ggplot object
ggplotObject <- ggplot(mtcars, aes(y=mpg, x=wt, color=as.factor(cyl))) +
    geom_point(size = 3, alpha = 0.5) +
    scale_color_cmac()

# Add the theme
ggplotObject + theme_cmac()

# Add the theme with additional optional theme parameters
ggplotObject + theme_cmac(legend.position = 'right')

}
\seealso{
\code{\link{scale_cmac}}

\code{\link{scale_color_cmac}}

\code{\link{scale_fill_cmac}}
}
